chore: Only clear if config says so.

+++ src/main/java/de/presti/smphelper/listener/ReadyListener.java
@@ -35,11 +35,13 @@ public class ReadyListener extends ListenerAdapter {
             }
         }
 
-        var category = event.getJDA().getCategoryById(Main.getTemporalVoiceCategory());
-        if (category != null) {
-            for (var channel : category.getVoiceChannels()) {
-                if (channel.getMembers().isEmpty()) {
-                    channel.delete().queue();
+        if (Config.getInstance().isClearTemporalVCsOnStart()) {
+            var category = event.getJDA().getCategoryById(Main.getTemporalVoiceCategory());
+            if (category != null) {
+                for (var channel : category.getVoiceChannels()) {
+                    if (channel.getMembers().isEmpty()) {
+                        channel.delete().queue();
+                    }
                 }
             }
         }
